Handle too large values for integers

+++ init_db.sql
@@ -1,7 +1,7 @@
 CREATE TABLE fizzbuzz_requests (
     id serial PRIMARY KEY,
-    int1 INTEGER NOT NULL,
-    int2 INTEGER NOT NULL,
+    int1 BIGINT NOT NULL,
+    int2 BIGINT NOT NULL,
     str1 VARCHAR (355) NOT NULL,
     str2 VARCHAR (355) NOT NULL
 )
